pyproject.toml
[tool.poetry]
name = "dnd"
version = "0.1.0"
description = ""
authors = ["AliiMustafa <aliimustafazada.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.38"
fastapi = {extras = ["all"], version = "^0.115.8"}
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
jose = "^1.0.0"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"
python-dotenv = "^1.0.1"
mysqlclient = "^2.2.7"
pymysql = "^1.1.1"
python-jose = "^3.4.0"
pytest = "^8.3.4"
aiomysql = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
